requirements.txt
mcstatus==4.1.0
discord.py==1.6.0
docker==4.4.1